Revert "Shrink more weight arrays to uint8_t"
This reverts commit 74625e9f47435736b37c504a355ec3e6bfa71917.

+++ Source/astcenc_block_sizes2.cpp
@@ -780,7 +780,7 @@ static void construct_block_size_descriptor_2d(
 		astc::rand_init(rng_state);
 
 		// pick 64 random texels for use with bitmap partitioning.
-		uint8_t arr[MAX_TEXELS_PER_BLOCK];
+		int arr[MAX_TEXELS_PER_BLOCK];
 		for (int i = 0; i < x_dim * y_dim; i++)
 		{
 			arr[i] = 0;
@@ -948,7 +948,7 @@ static void construct_block_size_descriptor_3d(
 		astc::rand_init(rng_state);
 
 		// pick 64 random texels for use with bitmap partitioning.
-		uint8_t arr[MAX_TEXELS_PER_BLOCK];
+		int arr[MAX_TEXELS_PER_BLOCK];
 		for (int i = 0; i < xdim * ydim * zdim; i++)
 		{
 			arr[i] = 0;
